Split IcebergHelper: detection vs RAPIDS GPU support gates
IcebergHelper had grown into a grab bag carrying three distinct
responsibilities: Iceberg app detection, Iceberg-specific table/exec metadata,
and RAPIDS GPU support gates. This commit splits the third concern into a new
IcebergGpuSupport object so each file owns a single role.

IcebergHelper now keeps:
- extensionRegxMap, eval, isSparkCatalogDefined, getCatalogType  (detection)
- ICEBERG_METADATA_TABLE_SUFFIXES                                (table metadata)
- EXEC_* constants, DEFINED_EXECS                                (exec stubs)

IcebergGpuSupport (new) owns:
- SUPPORTED_CATALOGS, SUPPORTED_SPARK_VERSION_REGEX, CONF_* constants
- isSparkVersionSupported, isIcebergFormatEnabled, isIcebergWriteEnabled,
  isParquetFieldIdWriteEnabled, isSparkCatalogSupported, isConfDisabled
- firstUnsupportedWritePrerequisite  (renamed from firstUnsupportedRuntimeReason
  to better reflect that it gates write-eligibility prerequisites, not a
  "runtime" notion)
- firstUnsupportedCatalogReason

Pure code movement; no behavior change. AppendDataIcebergParser,
MergeRowsIcebergParser, and BaseNoSparkSuite now point at IcebergGpuSupport for
the moved methods. IcebergGpuSupport.isSparkCatalogSupported delegates to
IcebergHelper.getCatalogType (which stays in the detection object since it is
also used by isSparkCatalogDefined / eval).

The IcebergHelperSuite test file was renamed to IcebergGpuSupportSuite since
every test in it now exercises IcebergGpuSupport methods. No test bodies
changed.

Deliberately skipped the parser-side trait (IcebergWriteGateChecks) that an
earlier design pass suggested: with only two parser call sites and a four-line
plumb-through each, a shared trait is premature abstraction. The duplication
diagnostic on AppendDataIcebergParser:101 / MergeRowsIcebergParser:131 is
accepted.

Tests: IcebergGpuSupportSuite 10/10, QualificationSuite 32/32, scalastyle
clean.
